hpc and comp. steps separated + ga core renaming
