+++ Green_Algorithms_core/src/ga_core/data_models/job_emissions_record.py
