+++ Green_Algorithms_core/src/ga_core/data_models/normalised_job_record.py
