+++ ga_dashboard/backend/ga_tools.py
@@ -1,5 +1,6 @@
 # ------------------------------------------------------------------
 # Main tools to extract, enrich and summarise data for the GA4HPC dashboard.
+# Uses GA Core for data extraction and processing
 # ------------------------------------------------------------------
 
 import pandas as pd
@@ -7,7 +8,7 @@ import yaml
 import time as time_module
 import gc
 from tqdm import tqdm
-import Green_Algorithms_core
+import Green_Algorithms_core.src.ga_core as ga_core
 
 from ga_dashboard.backend.services.database_ci_store import DatabaseCIStore
 import ga_dashboard.backend.helpers.utils as utils
@@ -163,7 +164,7 @@ class LogsDataProcessor:
         :return: pandas Dataframe containing processed data
         """
         db_ci_store = DatabaseCIStore(self.db_params)
-        dataprocessor = Green_Algorithms_core.HPCDataProcessor(self.config_data, self.cluster_info, self.fixed_params, self.has_slurmAdmin)
+        dataprocessor = ga_core.HPCDataProcessor(self.config_data, self.cluster_info, self.fixed_params, self.has_slurmAdmin)
         df = dataprocessor.extract_data()
         df = dataprocessor.enrich_data(df, db_ci_store)
