+++ ga_dashboard/backend/helpers/utils.py
@@ -11,29 +11,6 @@ import numpy as np
 from datetime import timedelta
 from io import BytesIO
 
-def check_empty_results(df, args):
-    """
-    This is to check whether any jobs have been run in the period, and stop the script if not.
-    :param df: [pd.DataFrame] Usage logs
-    :param args: [argStruct] Named tuple of arguments used.
-    """
-    if len(df) == 0:
-        if args.filterWD is not None:
-            addThat = f' from this directory ({args.filterWD})'
-        else:
-            addThat = ''
-        if args.filterJobIDs != 'all':
-            addThat += ' and with these jobIDs'
-        if args.filterAccount is not None:
-            addThat += ' charged under this account'
-
-        print(f'''
-
-    You haven't run any jobs in that period (from {args.startDay} to {args.endDay}){addThat}.
-
-        ''')
-        sys.exit()
-
 def parse_string_to_number(s:str) -> int | float | str:
     try:
         return int(s)
@@ -42,25 +19,6 @@ def parse_string_to_number(s:str) -> int | float | str:
             return float(s)
         except ValueError:
             return s
-        
-def convert2dataframe(df_raw: bytes, types: dict | None = None, delimiter="|"):
-    """
-    Convert raw logs output into a pandas DataFrame.
-    Parameters:
-        df_raw : Raw logs output as bytes.
-        types : column names and their desired data types. E.g., {'NNodes': 'int64', 'NCPUS': 'int64'}
-        delimiter : Delimiter used in the raw logs.
-    Returns:
-        pd.DataFrame: DataFrame containing the parsed logs with specified data types.
-    """
-    df = pd.read_csv(BytesIO(df_raw), sep=delimiter, dtype='str')
-
-    # Convert specified columns to appropriate data types 
-    if types:
-        for c, t in types.items():
-            if c in df.columns:
-                df[c] = df[c].astype(t)
-    return df
         
 def generate_batches_by_dates(start: str | datetime.date, end: str | datetime.date, batch_size: int = 30) -> list:
     """
@@ -90,19 +48,6 @@ def generate_batches_by_dates(start: str | datetime.date, end: str | datetime.da
         current_start = current_end + timedelta(days=1)
 
     return batches
-
-def concat_dataframes(dfs: list[pd.DataFrame]) -> pd.DataFrame:
-    """
-    Concatenate DataFrames after filtering out empty or all-NaN inputs.
-
-    This ensures consistent dtype inference and avoids future incompatibilities with pandas, 
-    where concatenation behavior with empty or all-NaN DataFrames is changing (FutureWarning).
-    """
-
-    # Keep only DataFrames that are not empty and not entirely NaN
-    dfs = [df for df in dfs if not df.empty and not df.isna().all().all()]
-    
-    return pd.concat(dfs, ignore_index=True)
 
 ##DEBUGONLY 
 def simulate_mock_jobs():
@@ -176,24 +121,5 @@ def quick_inspect(df: pd.DataFrame, name: str = "DataFrame") -> None:
     print("Head:\n", df.head())
     print(f"--- End of {name} Inspection ---\n")
 
-##DEBUGONLY 
-def get_mock_agg_data() -> pd.DataFrame:
-    """
-    Read and return mock aggregated data from a pickled file. Mock data generated using 'simulate_mock_jobs()' function.
-    """
-    # Steps done in pickle_it.py script:
-    # df2 = simulate_mock_jobs()
-    # df2.to_pickle("testdata/df_agg_X_mockMultiUsers_1.pkl")
-    # NB the data generated is different each time.
-
-    # foo = 'testdata/df_agg_test_3.pkl'
-    # foo = 'testdata/df_agg_X_1.pkl'
-
-    has_slurmAdmin = True # Assuming we have admin access
-    if has_slurmAdmin: 
-        pickled_test_data = 'tests/testdata/df_agg_X_mockMultiUsers_2.pkl'
-        
-    print(f"Overriding df_agg with `{pickled_test_data}`")
-    return pd.read_pickle(pickled_test_data)
